+++ libs/create-qwikdev-astro/src/utils.ts
@@ -1,11 +1,120 @@
-import fs from "node:fs";
+import fs, { readdirSync, statSync } from "node:fs";
 import os from "node:os";
 import path, { join, resolve, relative } from "node:path";
 import { fileURLToPath } from "node:url";
+import { copySync, ensureDirSync, pathExistsSync } from "fs-extra/esm";
 import detectPackageManager from "which-pm-runs";
 
 export const __filename = getModuleFilename();
 export const __dirname = path.dirname(__filename);
+
+export function safeCopy(source: string, target: string): void {
+  const files = readdirSync(source);
+
+  for (const file of files) {
+    const sourcePath = join(source, file);
+    const targetPath = join(target, file);
+
+    if (statSync(sourcePath).isDirectory()) {
+      ensureDirSync(targetPath);
+      safeCopy(sourcePath, targetPath);
+    } else if (!pathExistsSync(targetPath)) {
+      copySync(sourcePath, targetPath);
+    } else if (file.endsWith(".json")) {
+      deepMergeJsonFile(targetPath, sourcePath, true);
+    } else if (file.startsWith(".") && file.endsWith("ignore")) {
+      mergeDotIgnoreFiles(targetPath, sourcePath, true);
+    }
+  }
+}
+
+export function deepMergeJsonFile<T>(
+  targetJsonPath: string,
+  sourceJsonPath: string,
+  replace = false
+): T {
+  const deepMerge = deepMergeJson<T>(
+    fileGetContents(targetJsonPath),
+    fileGetContents(sourceJsonPath)
+  );
+
+  if (replace) {
+    putJson(targetJsonPath, deepMerge);
+  }
+
+  return deepMerge;
+}
+
+export function deepMergeJson<T>(targetJson: string, sourceJson: string): T {
+  return deepMerge(JSON.parse(targetJson), JSON.parse(sourceJson)) as unknown as T;
+}
+
+export function deepMerge<T>(target: T, source: Partial<T>): T {
+  for (const key of Object.keys(source) as (keyof T)[]) {
+    const targetValue = target[key];
+    const sourceValue = source[key] as Partial<T[keyof T]>;
+
+    if (isObject(targetValue) && isObject(sourceValue)) {
+      target[key] = deepMerge(targetValue, sourceValue);
+    } else if (Array.isArray(targetValue) && Array.isArray(sourceValue)) {
+      target[key] = Array.from(new Set([...targetValue, ...sourceValue])) as any;
+    } else {
+      target[key] = sourceValue as T[keyof T];
+    }
+  }
+  return target;
+}
+
+function isObject(item: unknown): item is Record<string, any> {
+  return item !== null && typeof item === "object" && !Array.isArray(item);
+}
+
+export function mergeDotIgnoreFiles(
+  target: string,
+  source: string,
+  replace = false
+): string {
+  const contents = mergeDotIgnoreContents(
+    fileGetContents(target),
+    fileGetContents(source)
+  );
+
+  if (replace) {
+    filePutContents(target, contents);
+  }
+
+  return contents;
+}
+
+export function mergeDotIgnoreContents(content1: string, content2: string): string {
+  return mergeDotIgnoreLines(content1.split("\n"), content2.split("\n")).join("\n");
+}
+
+export function mergeDotIgnoreLines(lines1: string[], lines2: string[]): string[] {
+  const lines = Array.from(
+    new Set([...lines1.map((line) => line.trim()), ...lines2.map((line) => line.trim())])
+  ).filter((line) => line !== "");
+
+  return formatLines(lines);
+}
+
+function formatLines(lines: string[]): string[] {
+  const formattedLines: string[] = [];
+  let previousWasComment = false;
+
+  lines.forEach((line, index) => {
+    const isComment = line.startsWith("#");
+
+    if (isComment && !previousWasComment && index !== 0) {
+      formattedLines.push("");
+    }
+
+    formattedLines.push(line);
+    previousWasComment = isComment;
+  });
+
+  return formattedLines;
+}
 
 export function getModuleFilename(): string {
   const error = new Error();
@@ -140,7 +249,11 @@ export function getPackageJson(dir: string): Record<string, any> {
 }
 
 export function setPackageJson(dir: string, json: Record<string, any>) {
-  filePutContents(getPackageJsonPath(dir), JSON.stringify(json, null, 2));
+  putJson(getPackageJsonPath(dir), json);
+}
+
+export function putJson<T>(path: string, json: T) {
+  filePutContents(path, JSON.stringify(json, null, 2));
 }
 
 export function updatePackageName(newName: string, dir = __dirname): void {