+++ libs/qwikdev-astro/src/utils.ts
@@ -1,5 +1,0 @@
-export function newHash() {
-  const hash = Math.random().toString(26).split(".").pop();
-  globalThis.hash = hash;
-  return hash;
-}