+++ libs/qwikdev-astro/src/virtual.d.ts
@@ -1,0 +1,4 @@
+declare module "inox:inline-mod:mod_0" {
+  export const isNode: boolean;
+  export const qAstroManifestPath: string;
+}